requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
PyMySQL==1.1.0
mysqlclient==2.2.0
pyrebase4==4.7.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
SQLAlchemy==2.0.23
WTForms==3.1.0
email-validator==2.1.0
firebase-admin==6.2.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0
